requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
gTTS
g4f
